+++ build.rs
@@ -1,7 +1,7 @@
 extern crate winres;
 
 fn main() {
-    println!("sixtyfps build");
+    println!("slint  build");
     slint_build::compile("ui/image_sieve.slint").unwrap();
     if cfg!(target_os = "windows") {
         let mut res = winres::WindowsResource::new();