+++ src/persistence/model_to_enum.rs
@@ -1,7 +1,7 @@
 use num_traits::{FromPrimitive, ToPrimitive};
 use slint::{Model, ModelRc, SharedString};
 
-/// Convert a value from a sixtyfps SharedString model to an enum by mapping the model index to the enum value.
+/// Convert a value from a slint SharedString model to an enum by mapping the model index to the enum value.
 pub fn model_to_enum<Enum>(model: &ModelRc<SharedString>, value: &SharedString) -> Enum
 where
     Enum: FromPrimitive,
@@ -15,7 +15,7 @@ where
     enum_value
 }
 
-/// Convert an enum to a value from a sixtyfps SharedString modelby mapping the enum value to the model index.
+/// Convert an enum to a value from a slint SharedString modelby mapping the enum value to the model index.
 pub fn enum_to_model<Enum>(model: &ModelRc<SharedString>, value: &Enum) -> SharedString
 where
     Enum: ToPrimitive,